backend/requirements.txt
psycopg2==2.9.*
Django==4.1.*
djangorestframework==3.14.*
hashids==1.3.*
django-cors-headers==3.13.*